pyproject.toml
[tool.poetry]
name = "signate1106"
version = "0.1.0"
description = "signate1106"
authors = [
	"AOYAMA hiroyuki <3ign0n@gmail.com>"
]

[tool.kedro]
package_name = "signate1106"
project_name = "signate1106"
kedro_init_version = "0.18.11"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
black = ">=22.0,<23.0"
flake8 = ">=3.7.9,<5.0"
ipython = {version = ">=8.10,<9.0", markers = "python_version >= \"3.8\""}
isort = ">=5.0,<6.0"
jupyter = ">=1.0,<2.0"
jupyterlab-server = ">=2.11.1,<2.16.0"
jupyterlab = ">=3.0,<3.6.0"
kedro-telemetry = ">=0.2.0,<0.3.0"
nbstripout = ">=0.4,<1.0"
pytest-cov = ">=3.0,<4.0"
pytest-mock = ">=1.7.1,<2.0"
pytest = ">=7.2,<8.0"
pandas = "^2.0.3"
kedro-datasets = {version = ">=1.0,<2.0", extras = ["pandas-csvdataset", "pandas-parquetdataset"]}
kedro-mlflow = "^0.11.9"
optuna = "^3.2.0"
lightgbm = "^4.0.0"
category-encoders = "^2.6.1"
matplotlib = "^3.7.2"
pyqt5 = "^5.15.9"
pandas-profiling = "^3.6.6"
plotly-express = "^0.4.1"
kaleido = "0.2.1"
pillow = "^10.0.0"
plotly = "^5.15.0"
torch = {url = "https://download.pytorch.org/whl/cpu/torch-2.0.1-cp39-none-macosx_10_9_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cpu/torchvision-0.15.2-cp39-cp39-macosx_10_9_x86_64.whl"}
sympy = {url = "https://download.pytorch.org/whl/sympy-1.11.1-py3-none-any.whl"}
mpmath = {url = "https://download.pytorch.org/whl/mpmath-1.2.1-py3-none-any.whl"}
filelock = {url = "https://download.pytorch.org/whl/filelock-3.9.0-py3-none-any.whl"}

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/signate1106 -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]
